// src/domain/enums.ts
export enum ControlCommand {
  stop = 0x10,
  up = 0x20,
  ventilationPosTilting = 0x24,
  down = 0x40,
  intermediatePos = 0x44,
}

export enum EasyCommand {
  EASY_CHECK = 0x4a,
  EASY_CONFIRM = 0x4b,
  EASY_SEND = 0x4c,
  EASY_ACK = 0x4d,
  EASY_INFO = 0x4e,
}

export enum InfoData {
  INFO_UNKNOWN = -1,
  INFO_NO_INFORMATION = 0x00,
  INFO_TOP_POSITION_STOP = 0x01,
  INFO_BOTTOM_POSITION_STOP = 0x02,
  INFO_INTERMEDIATE_POSITION_STOP = 0x03,
  INFO_TILT_VENTILATION_POS_STOP = 0x04,
  INFO_BLOCKING = 0x05,
  INFO_OVERHEATED = 0x06,
  INFO_TIMEOUT = 0x07,
  INFO_START_TO_MOVE_UP = 0x08,
  INFO_START_TO_MOVE_DOWN = 0x09,
  INFO_MOVING_UP = 0x0a,
  INFO_MOVING_DOWN = 0x0b,
  INFO_STOPPED_IN_UNDEFINED_POSITION = 0x0d,
  INFO_TOP_POS_STOP_WICH_TILT_POS = 0x0e,
  INFO_BOTTOM_POS_STOP_WICH_INT_POS = 0x0f,
  INFO_SWITCHING_DEVICE_SWITCHED_OFF = 0x10,
  INFO_SWITCHING_DEVICE_SWITCHED_ON = 0x11,
}